move everything out of utils into ssr.js or default.js
